pyproject.toml
[project]
name = "wh-foundry"
version = "0.1.0"
description = "Western Health knowledge base scraper and Azure AI Search tooling"
requires-python = ">=3.12"

dependencies = [
    "azure-ai-projects>=2.0",
    "azure-ai-voicelive>=1.0",
    "azure-identity>=1.25",
    "azure-storage-blob>=12.28",
    "beautifulsoup4>=4.14",
    "markdownify>=1.2",
    "playwright>=1.58",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
]

[tool.setuptools]
py-modules = ["scrape_pages", "upload_docs", "chat"]

[project.scripts]
scrape-pages = "scrape_pages:main"
upload-docs = "upload_docs:main"
